refactor: extract ingredient usage calculation to domain layer use case
Move ~70 lines of business logic from RecipeDetailViewModel to
CalculateIngredientUsageUseCase. This includes ingredient total
computation, used amount accumulation from checked instruction steps,
and final usage status determination.

Also moves IngredientUsageStatus, InstructionIngredientKey, and
createInstructionIngredientKey to domain model layer since they
represent domain concepts used by the use case.

Closes #72
